+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
-- Fix: Excel Import should return ODCS YAML (#829)
+- Excel Import should return ODCS YAML (#829)
+- Excel Import: Missing server section when the server included a schema property (#823)
 
 ### Changed
 - Use `&#x2007;` instead of `&numsp;` for tab in Markdown export.